move lipsum util to stories folder
