+++ src/stories/utils.js
@@ -1,0 +1,3 @@
+import { loremIpsum } from 'lorem-ipsum'
+
+export let lipsum = (count, units) => loremIpsum({ count, units })
